add product filter, pagination, implement user model, user controller, routes, implement register, login, set JWT token

+++ controllers/productController.js
@@ -1,6 +1,7 @@
 const Product = require("../models/productModel");
 const ErrorHandler = require("../utils/errorHandler");
 const catchAsyncErrors = require("../middleware/catchAsyncErrors");
+const ApiFeatures = require("../utils/apiFeatures");
 
 
 //Create product
@@ -11,19 +12,35 @@ exports.createProduct = catchAsyncErrors(async (req,res,next)=>{
 
     res.status(201).json({
         success:true,
-        product,
+        message:`Product ${product.name} created successfully`,
+
     })
 });
 
 //Get All Product
-exports.getAllProduct= catchAsyncErrors(async (req,res)=>{
-    const products = await Product.find();
+exports.getAllProduct = catchAsyncErrors(async (req, res) => {
+
+const resultPerPage = 5;
+
+
+
+    const apiFeatures = new ApiFeatures(Product.find(), req.query)
+        .search()
+          // chain search
+        .filter()
+        // chain filter
+        .pagination(resultPerPage); // 10 products per page
+
+    const products = await apiFeatures.query; // ✅ execute the mongoose query here
+
     res.status(200).json({
-        success:true,
-        message:"fetch succesfully",
-        products,
-    })
+        success: true,
+        message: "Fetch successfully",
+        count: products.length,
+        products,   // now it's plain documents
+    });
 });
+
 
 //get product detail
 exports.getProductDetail = catchAsyncErrors(async (req,res,next) =>{
@@ -58,13 +75,19 @@ res.status(200).json({
 
 
 //delete product
-exports.deleteProduct = catchAsyncErrors(async (req,res,next)=> {
-    const product =await Product.findById(req.params.id);
-  return next(new ErrorHandler("Product Not Found",404));``
+exports.deleteProduct = catchAsyncErrors(async (req,res,next) => {
+
+    const product = await Product.findById(req.params.id);
+
+    if(!product) {
+         return next(new ErrorHandler("Product Not Found",404));
+    }
+
     await product.deleteOne();
 
     res.status(200).json({
         success:true,
-        message:"product deleted"
-    })
+        message:` ${product.name} deleted`
+    })      
 });
+
